# Migrate `State.Context` API to inline functions