--- a/library/keccak/api/keccak.klib.api
+++ b/library/keccak/api/keccak.klib.api
@@ -40,6 +40,7 @@ sealed class <#A: kotlin/Number, #B: org.kotlincrypto.sponges.keccak/State<#A, #
     final val state // org.kotlincrypto.sponges.keccak/State.state|{}state[0]
         final fun <get-state>(): kotlin/Array<#A> // org.kotlincrypto.sponges.keccak/State.state.<get-state>|<get-state>(){}[0]
 
+    abstract fun (#A).mixIn(#A): #A // org.kotlincrypto.sponges.keccak/State.mixIn|mixIn@1:0(1:0){}[0]
     abstract fun copy(): #B // org.kotlincrypto.sponges.keccak/State.copy|copy(){}[0]
     final fun addData(kotlin/Int, #A) // org.kotlincrypto.sponges.keccak/State.addData|addData(kotlin.Int;1:0){}[0]
     final fun contains(#A): kotlin/Boolean // org.kotlincrypto.sponges.keccak/State.contains|contains(1:0){}[0]
@@ -52,4 +53,8 @@ sealed class <#A: kotlin/Number, #B: org.kotlincrypto.sponges.keccak/State<#A, #
     final object Companion // org.kotlincrypto.sponges.keccak/State.Companion|null[0]
 }
 
+final fun (org.kotlincrypto.sponges.keccak/F1600).org.kotlincrypto.sponges.keccak/keccakP(kotlin/Byte =...) // org.kotlincrypto.sponges.keccak/keccakP|keccakP@org.kotlincrypto.sponges.keccak.F1600(kotlin.Byte){}[0]
+final fun (org.kotlincrypto.sponges.keccak/F200).org.kotlincrypto.sponges.keccak/keccakP(kotlin/Byte =...) // org.kotlincrypto.sponges.keccak/keccakP|keccakP@org.kotlincrypto.sponges.keccak.F200(kotlin.Byte){}[0]
+final fun (org.kotlincrypto.sponges.keccak/F400).org.kotlincrypto.sponges.keccak/keccakP(kotlin/Byte =...) // org.kotlincrypto.sponges.keccak/keccakP|keccakP@org.kotlincrypto.sponges.keccak.F400(kotlin.Byte){}[0]
+final fun (org.kotlincrypto.sponges.keccak/F800).org.kotlincrypto.sponges.keccak/keccakP(kotlin/Byte =...) // org.kotlincrypto.sponges.keccak/keccakP|keccakP@org.kotlincrypto.sponges.keccak.F800(kotlin.Byte){}[0]
 final fun <#A: kotlin/Number> org.kotlincrypto.sponges.keccak/KeccakP(org.kotlincrypto.sponges.keccak/State<#A, *>, kotlin/Byte =...) // org.kotlincrypto.sponges.keccak/KeccakP|KeccakP(org.kotlincrypto.sponges.keccak.State<0:0,*>;kotlin.Byte){0§<kotlin.Number>}[0]