--- a/library/keccak/api/keccak.klib.api
+++ b/library/keccak/api/keccak.klib.api
@@ -40,7 +40,8 @@ sealed class <#A: kotlin/Number, #B: org.kotlincrypto.sponges.keccak/State<#A, #
     final val state // org.kotlincrypto.sponges.keccak/State.state|{}state[0]
         final fun <get-state>(): kotlin/Array<#A> // org.kotlincrypto.sponges.keccak/State.state.<get-state>|<get-state>(){}[0]
 
-    abstract fun (#A).mixIn(#A): #A // org.kotlincrypto.sponges.keccak/State.mixIn|mixIn@1:0(1:0){}[0]
+    abstract fun (#A).XOR(#A): #A // org.kotlincrypto.sponges.keccak/State.XOR|XOR@1:0(1:0){}[0]
+    abstract fun (kotlin/Long).toN(): #A // org.kotlincrypto.sponges.keccak/State.toN|toN@kotlin.Long(){}[0]
     abstract fun copy(): #B // org.kotlincrypto.sponges.keccak/State.copy|copy(){}[0]
     final fun addData(kotlin/Int, #A) // org.kotlincrypto.sponges.keccak/State.addData|addData(kotlin.Int;1:0){}[0]
     final fun contains(#A): kotlin/Boolean // org.kotlincrypto.sponges.keccak/State.contains|contains(1:0){}[0]