--- a/library/keccak/src/commonMain/kotlin/org/kotlincrypto/sponges/keccak/F200.kt
+++ b/library/keccak/src/commonMain/kotlin/org/kotlincrypto/sponges/keccak/F200.kt
@@ -15,8 +15,6 @@
  **/
 package org.kotlincrypto.sponges.keccak
 
-import kotlin.experimental.and
-import kotlin.experimental.inv
 import kotlin.experimental.xor
 import kotlin.jvm.JvmSynthetic
 
@@ -27,22 +25,7 @@ public class F200: State<Byte, F200> {
     public constructor(): super(roundCount = 18, state = Array(P_LEN) { 0 })
     private constructor(state: F200): super(state.roundCount, state.state.copyOf())
     public override fun copy(): F200 = F200(this)
-
+    protected override fun Byte.mixIn(data: Byte): Byte = this xor data
     @JvmSynthetic
-    internal override fun <T: Any?> withContext(block: Context<Byte>.() -> T): T = block(F200Context)
-
-    private object F200Context: Context<Byte> {
-        private const val MASK: Int = 0xff
-
-        override fun and(a: Byte, other: Byte): Byte = a and other
-        override fun inv(a: Byte): Byte = a.inv()
-        override fun xor(a: Byte, other: Byte): Byte = a xor other
-        override fun rotateLeft(a: Byte, n: Int): Byte {
-            val bitCount = n % Byte.SIZE_BITS
-            return (
-                ((a.toInt() and MASK) shl bitCount) or ((a.toInt() and MASK) ushr (Byte.SIZE_BITS - bitCount))
-            ).toByte()
-        }
-        override fun RC(index: Int): Byte = RC[index].toByte()
-    }
+    internal override fun RC(index: Int): Byte = RC[index].toByte()
 }