--- a/library/keccak/src/commonMain/kotlin/org/kotlincrypto/sponges/keccak/F800.kt
+++ b/library/keccak/src/commonMain/kotlin/org/kotlincrypto/sponges/keccak/F800.kt
@@ -15,8 +15,6 @@
  **/
 package org.kotlincrypto.sponges.keccak
 
-import kotlin.jvm.JvmSynthetic
-
 /**
  * [State] for Keccak-f[800]
  * */
@@ -24,7 +22,6 @@ public class F800: State<Int, F800> {
     public constructor(): super(roundCount = 22, state = Array(P_LEN) { 0 })
     private constructor(state: F800): super(state.roundCount, state.state.copyOf())
     public override fun copy(): F800 = F800(this)
-    protected override fun Int.mixIn(data: Int): Int = this xor data
-    @JvmSynthetic
-    internal override fun RC(index: Int): Int = RC[index].toInt()
+    protected override fun Int.XOR(data: Int): Int = this xor data
+    protected override fun Long.toN(): Int = toInt()
 }