--- a/library/keccak/src/commonTest/kotlin/org/kotlincrypto/sponges/keccak/KeccakPUnitTest.kt
+++ b/library/keccak/src/commonTest/kotlin/org/kotlincrypto/sponges/keccak/KeccakPUnitTest.kt
@@ -21,6 +21,7 @@ import org.kotlincrypto.endians.BigEndian
 import kotlin.test.Test
 import kotlin.test.assertContentEquals
 
+@Suppress("DEPRECATION")
 class KeccakPUnitTest {
 
     private fun <N: Number> assertKeccakP(state: State<N, *>, permutation1: List<N>, permutation2: List<N>) {