Closes #18