@05nelsonm 2024-12-18
Before/After benchmark results

<details>
    <summary>Linux</summary>

Before
```
jvm summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  22.767 ± 0.101  us/op

nativeHost summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  60.800 ± 0.281  us/op

js summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  46.880 ± 0.148  us/op

wasmJs summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  28.808 ± 0.073  us/op
```

After
```
jvm summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  15.535 ± 0.080  us/op

nativeHost summary:
Benchmark                 Mode  Cnt  Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  0.916 ± 0.002  us/op

js summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  38.207 ± 0.113  us/op

wasmJs summary:
Benchmark                 Mode  Cnt  Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  1.043 ± 0.004  us/op
```

</details>

<details>
    <summary>macOS</summary>

Before
```
jvm summary:
Benchmark                 Mode  Cnt   Score    Error  Units
KeccakPBenchmark.keccakP  avgt    5  35.286 ± 21.767  us/op

nativeHost summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  49.265 ± 2.805  us/op

js summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  44.664 ± 0.852  us/op

wasmJs summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  20.889 ± 1.221  us/op
```

After
```
jvm summary:
Benchmark                 Mode  Cnt   Score    Error  Units
KeccakPBenchmark.keccakP  avgt    5  19.974 ± 20.496  us/op

nativeHost summary:
Benchmark                 Mode  Cnt  Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  0.646 ± 0.065  us/op

js summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  39.706 ± 2.310  us/op

wasmJs summary:
Benchmark                 Mode  Cnt  Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  0.656 ± 0.035  us/op
```

</details>

<details>
    <summary>Windows</summary>

Before
```
jvm summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  24.383 � 4.257  us/op

nativeHost summary:
Benchmark                 Mode  Cnt    Score    Error  Units
KeccakPBenchmark.keccakP  avgt    5  122.284 ± 19.196  us/op

js summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  46.810 ± 0.164  us/op

wasmJs summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  27.840 ± 0.139  us/op
```

After
```
jvm summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  17.503 � 7.032  us/op

nativeHost summary:
Benchmark                 Mode  Cnt  Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  1.472 ± 0.005  us/op

js summary:
Benchmark                 Mode  Cnt   Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  38.398 ± 0.049  us/op

wasmJs summary:
Benchmark                 Mode  Cnt  Score   Error  Units
KeccakPBenchmark.keccakP  avgt    5  1.045 ± 0.003  us/op
```

</details>